+++ S1/C++/TP1/path.cc
@@ -7,8 +7,8 @@ Path::Path(int n, std::size_t w,std::size_t h){
 	hmax = h;
 	wmax = w;
 	my_path = new int[n];
-	int current_x = origin[0];
-	int current_y = origin[1];
+	size_t current_x = origin[0];
+	size_t current_y = origin[1];
 	int i = 0;
 	while(i<n){
 		int dir = rand()%8;
@@ -20,12 +20,12 @@ Path::Path(int n, std::size_t w,std::size_t h){
 			i++;
 		}
 	}
-	
+
 }
 void Path::print() const{
 	std::cout << '(' << origin[0]<<',';
 	std::cout << origin[1] << ")\t";
-	for (size_t i=0;i<length-1;i++){
+	for (int i=0;i<length-1;i++){
 		std::cout << direction_human[i] << '-';
 	}
 	std::cout << direction_human[length]<<std::endl;
